Time: 161 ms (37.4%), Space: 151.1 MB (52.28%) - LeetHub

+++ 3524-find-x-value-of-array-i/3524-find-x-value-of-array-i.cpp
@@ -1,3 +1,4 @@
+
 class Solution {
 public:
     vector<long long> resultArray(vector<int>& nums, int k) {
@@ -7,8 +8,8 @@ public:
         // whose product % k == r
         vector<long long> dp(k, 0);
 
-        for (int num : nums) {
-            int mod = num % k;
+        for (long long num : nums) {
+            long long mod = num % k;
 
             vector<long long> newDp(k, 0);
 
@@ -16,13 +17,13 @@ public:
             newDp[mod]++;
 
             // Extend all previous subarrays
-            for (int r = 0; r < k; r++) {
-                int newR = (r * mod) % k;
+            for (long long r = 0; r < k; r++) {
+                long long newR = (r * mod) % k;
                 newDp[newR] += dp[r];
             }
 
             // Add current subarrays to answer
-            for (int r = 0; r < k; r++) {
+            for (long long r = 0; r < k; r++) {
                 ans[r] += newDp[r];
             }
 
@@ -32,3 +33,35 @@ public:
         return ans;
     }
 };
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
